# .github/openapi.yml
openapi: "3.0.3"
info:
  title: TW Tournament App API
  version: "1.0.0"
servers:
  - url: http://localhost/api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
...
paths:
...
  /user/exists:
    get:
      summary: Check if a username or email is taken
      operationId: userExists
      parameters:
        - name: identifier
          in: query
          required: true
          schema:
            type: string
            example: player1
      responses:
        "200":
          description: Existence check result
        "400":
          description: Missing query param
...
  /user/update-username:
    post:
      summary: Update username (authenticated)
      operationId: updateUsername
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        "200":
          description: Username updated
        "401":
          description: Not authenticated
        "409":
          description: Username taken
  /user/update-password:
    post:
      summary: Update password (authenticated)
      operationId: updatePassword
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Password updated
        "401":
          description: Not authenticated or wrong current password
  /user/account:
    delete:
      summary: Delete own account (authenticated)
      operationId: deleteAccount
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Account deleted
        "401":
          description: Not authenticated
        "403":
          description: CSRF validation failed
  /user/stats:
    get:
      summary: Get stats for authenticated user
      operationId: getUserStats
      security:
        - sessionCookie: []
      responses:
        "200":
          description: User stats
        "401":
          description: Not authenticated

  # ── Guest ─────────────────────────────────────────────────────────────────
...
  /guest/username:
    post:
      summary: Set guest username
      operationId: updateGuestUsername
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        "200":
          description: Username set
        "401":
          description: Not a guest session

  # ── Password Reset ─────────────────────────────────────────────────────────
...
  /password-reset/verify:
    post:
      summary: Verify a password reset token
      operationId: verifyResetToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Token valid
        "400":
          description: Invalid or expired token
  /password-reset/reset:
    post:
      summary: Reset password using a valid token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Password reset
        "400":
          description: Invalid or expired token

  # ── Tournament ─────────────────────────────────────────────────────────────
...